+++ ndk/src/events/kinds/simple-group/metadata.ts
