+++ ndk/src/index.ts
@@ -26,7 +26,7 @@ export * from "./nwc/index.js";
 export * from "./thread/index.js";
 
 export * from "./events/kinds/simple-group/index.js";
-export * from "./events/kinds/simple-group/meta-data.js";
+export * from "./events/kinds/simple-group/metadata.js";
 export * from "./events/kinds/simple-group/member-list.js";
 
 export * from "./relay/index.js";
